requirements.txt
attrs==22.2.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
iniconfig==2.0.0
numpy==1.24.2
opencv-python==4.7.0.68
packaging==23.0
paramiko==3.0.0
Pint==0.20.1
pluggy==1.0.0
pycparser==2.21
pyfrc==2023.0.1
PyNaCl==1.5.0
pynetconsole==2.0.4
pynetworktables==2021.0.0
pyntcore==2023.3.2.1
pytest==7.2.1
pytest-reraise==2.1.2
robotpy==2023.3.2.3
robotpy-apriltag==2023.3.2.1
robotpy-hal==2023.3.2.1
robotpy-halsim-gui==2023.3.2.1
robotpy-installer==2023.0.2
robotpy-wpilib-utilities==2023.1.0
robotpy-wpimath==2023.3.2.1
robotpy-wpinet==2023.3.2.1
robotpy-wpiutil==2023.3.2.1
six==1.16.0
wpilib==2023.3.2.1
